requirements.txt
wandb
pytest
transformers
datasets
soundfile
simple-parsing
pytorch
pytorch-lightning